+++ src/main/kotlin/com/vermouthx/stocker/actions/StockerStockSearchAction.kt
@@ -6,7 +6,11 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.vermouthx.stocker.StockerBundle
 import com.vermouthx.stocker.views.dialogs.StockerSuggestionDialog
 
-class StockerStockSearchAction : AnAction() {
+class StockerStockSearchAction : AnAction(
+    StockerBundle.msg("action.add.favorite.stocks"),
+    StockerBundle.msg("action.add.favorite.stocks.description"),
+    null
+) {
     override fun update(e: AnActionEvent) {
         val project = e.project
         val presentation = e.presentation